+++ files_rpi/tmp/home/tty2rpi/tty2rpi-scripts/CMDSHOWIPS.sh
@@ -1,29 +1,21 @@
 #!/bin/bash
 
-source ~/tty2rpi.ini
-source ~/tty2rpi-user.ini
-source ~/tty2rpi-screens.ini
+. ~/tty2rpi.ini
+. ~/tty2rpi-user.ini
+. ~/tty2rpi-screens.ini
+. ~/tty2rpi-functions.ini
+#. <(cat ~/tty2rpi*.ini)
 
-${IMconvert} -size ${RESOLUTION} xc:black -pointsize $((${WIDTH}/20)) -fill white -gravity center -draw "text 0,0 'Waiting for network...'" /tmp/wait4inet.png
-if [ "${FBUFDEV}" = "yes" ]; then
-  FRAMEBUFFER="${FBDEVICE}" fim --autozoom --quiet --output-device fb /tmp/wait4inet.png > /dev/null 2>&1 &
-else
-  feh --quiet --fullscreen --auto-zoom /tmp/wait4inet.png &
-fi
+${IMconvert} -size ${RESOLUTION} xc:black -pointsize $((${WIDTH}/20)) -fill white -gravity center -draw "text 0,0 'Waiting for network...'" "${TMPDIR}/pictmp.png"
+mv "${TMPDIR}/pictmp.png" "${TMPDIR}/pic.png"
+until [ $(ip link | grep -c "state UP") -gt 0 ]; do true; sleep 0.2; done
 
-DEFGW="$(ip route | grep "default via" | awk '{print $3}')"
+DEFGW="$(ip route | head -n1 | grep "default via" | awk '{print $3}')"
 until ping -c1 "${DEFGW}" >/dev/null 2>&1; do sleep 1; done
-if [ "${FBUFDEV}" = "yes" ]; then
-  KILLPID fim
-else
-  KILLPID feh
-fi
 
-echo -e "Your IP addresss(es) for\nhostname $(hostname -f):\n" > /tmp/ip-addressses
-ip -o addr | awk '!/^[0-9]*: ?lo|link\/ether/ {gsub("/", " "); print $2" "$4}' >> /tmp/ip-addressses
-${IMconvert} -size ${RESOLUTION} xc:black -pointsize $((${WIDTH}/25)) -fill white -gravity center -draw "text 0,0 '$(</tmp/ip-addressses)'" /tmp/ip-addressses.png
-if [ "${FBUFDEV}" = "yes" ]; then
-  FRAMEBUFFER="${FBDEVICE}" fim --autozoom --quiet --output-device fb /tmp/ip-addressses.png > /dev/null 2>&1
-else
-  feh --quiet --fullscreen --auto-zoom /tmp/ip-addressses.png
-fi
+echo -e "Your IP address(es) for\nhostname $(hostname -f):\n" > /tmp/ip-addresses
+ip -o addr | awk '!/^[0-9]*: ?lo|link\/ether/ {gsub("/", " "); print $2" "$4}' >> /tmp/ip-addresses
+sed -i '/ fd23:/d' /tmp/ip-addresses
+sed -i '/ fe80:/d' /tmp/ip-addresses
+${IMconvert} -size ${RESOLUTION} xc:black -pointsize $((${WIDTH}/25)) -fill white -gravity center -draw "text 0,0 '$(</tmp/ip-addresses)'" "${TMPDIR}/pictmp.png"
+mv "${TMPDIR}/pictmp.png" "${TMPDIR}/pic.png"