+++ files_rpi/tmp/home/tty2rpi/tty2rpi-scripts/CMDSHUTDOWN.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
-source ~/tty2rpi.ini
-source ~/tty2rpi-user.ini
-source ~/tty2rpi-screens.ini
+. ~/tty2rpi.ini
+. ~/tty2rpi-user.ini
+. ~/tty2rpi-screens.ini
+. ~/tty2rpi-functions.ini
+#. <(cat ~/tty2rpi*.ini)
 
 # wget -q 'http://tasmota-012345-1234.example.,org/cm?cmnd=Backlog%3BDelay%20150%3BPower%20Off' -O /dev/null # tty2rpi
 sudo systemctl poweroff