+++ files_rpi/tmp/home/tty2rpi/tty2rpi-scripts/CMDSLIDESHOW.sh
@@ -1,14 +1,12 @@
 #!/bin/bash
 
-source ~/tty2rpi.ini
-source ~/tty2rpi-user.ini
-source ~/tty2rpi-screens.ini
+. ~/tty2rpi.ini
+. ~/tty2rpi-user.ini
+. ~/tty2rpi-screens.ini
+. ~/tty2rpi-functions.ini
+#. <(cat ~/tty2rpi*.ini)
 
 [ "${SCREENSAVER}" = "yes" ] && systemctl --user stop --quiet tty2rpi-screensaver.timer
 SHOWTOUT="${COMMANDLINE[1]}"
 [ -z "${SHOWTOUT}" ] && SHOWTOUT=9
-if [ "${FBUFDEV}" = "yes" ]; then
-  FRAMEBUFFER="${FBDEVICE}" fim --autozoom --quiet --output-device fb --slideshow ${SHOWTOUT} ${PATHPIC} > /dev/null 2>&1
-else
-  feh --quiet --fullscreen --auto-zoom --randomize --auto-zoom --slideshow-delay ${SHOWTOUT} ${PATHPIC}
-fi
+feh --quiet --fullscreen --auto-zoom --randomize --auto-zoom --slideshow-delay "${SHOWTOUT}" "${PATHPIC}"