+++ files_rpi/tmp/home/tty2rpi/tty2rpi-scripts/CMDSNAM.sh
@@ -1,13 +1,10 @@
 #!/bin/bash
 
-source ~/tty2rpi.ini
-source ~/tty2rpi-user.ini
-source ~/tty2rpi-screens.ini
+. ~/tty2rpi.ini
+. ~/tty2rpi-user.ini
+. ~/tty2rpi-screens.ini
+. ~/tty2rpi-functions.ini
+#. <(cat ~/tty2rpi*.ini)
 
-${IMconvert} -undercolor Black -fill white -pointsize $((${WIDTH}/20)) -gravity South -draw "text 0,0 '$(</dev/shm/corename)'" /dev/shm/pic.png /dev/shm/pic.png
-
-if [ "${FBUFDEV}" = "yes" ]; then
-  FRAMEBUFFER="${FBDEVICE}" fim --autozoom --quiet --output-device fb /dev/shm/pic.png > /dev/null 2>&1
-else
-  feh --quiet --fullscreen --auto-zoom /dev/shm/pic.png
-fi
+${IMconvert} "${TMPDIR}/pic.png" -undercolor Black -fill white -pointsize $((${WIDTH}/20)) -gravity South -draw "text 0,0 '$(<${TMPDIR}/tmp/corename)'" "${TMPDIR}/tmp/pictmp.png"
+mv "${TMPDIR}/tmp/pictmp.png" "${TMPDIR}/pic.png"