+++ files_rpi/tmp/home/tty2rpi/tty2rpi-scripts/CMDTEST.sh
@@ -1,26 +1,14 @@
 #!/bin/bash
 
-source ~/tty2rpi.ini
-source ~/tty2rpi-user.ini
-source ~/tty2rpi-screens.ini
+. ~/tty2rpi.ini
+. ~/tty2rpi-user.ini
+. ~/tty2rpi-screens.ini
+. ~/tty2rpi-functions.ini
+#. <(cat ~/tty2rpi*.ini)
 
 [ "${SCREENSAVER}" = "yes" ] && systemctl --user stop --quiet tty2rpi-screensaver.timer
 if [ -f "${PATHPIC}/000-TESTBILD.png" ]; then
-  PICSIZE=$(identify -format '%wx%h' "${PATHPIC}/${MEDIA}")
-  if [ "${PICSIZE}" != "${WIDTH}x${HEIGHT}" ]; then
-    ffmpeg -y -loglevel quiet -i "${PATHPIC}/000-TESTBILD.png" -vf scale=${WIDTH}:${HEIGHT} /tmp/testpic.png
-  else
-    if [ "${FBUFDEV}" = "yes" ]; then
-      FRAMEBUFFER="${FBDEVICE}" fim --autozoom --quiet --output-device fb "${PATHPIC}/000-TESTBILD.png" > /dev/null 2>&1
-    else
-      feh --quiet --fullscreen --auto-zoom "${PATHPIC}/000-TESTBILD.png"
-    fi
-  fi
+  cp "${PATHPIC}/000-TESTBILD.png" "${TMPDIR}/pic.png"
 else
-  cp ${TTY2RPIPICS}/000-notavail.png /dev/shm/pic.png
-fi
-if [ "${FBUFDEV}" = "yes" ]; then
-  FRAMEBUFFER="${FBDEVICE}" fim --autozoom --quiet --output-device fb /tmp/testpic.png > /dev/null 2>&1
-else
-  feh --quiet --fullscreen --auto-zoom /tmp/testpic.png
+  cp "${TTY2RPIPICS}/000-notavail.png" "${TMPDIR}/pic.png"
 fi