+++ files_rpi/tmp/home/tty2rpi/tty2rpi-scripts/tty2rpi-screensaver.sh
@@ -1,17 +1,15 @@
 #!/bin/bash
 
-source ~/tty2rpi.ini
-source ~/tty2rpi-user.ini
-source ~/tty2rpi-screens.ini
+. ~/tty2rpi.ini
+. ~/tty2rpi-user.ini
+. ~/tty2rpi-screens.ini
+. ~/tty2rpi-functions.ini
+#. <(cat ~/tty2rpi*.ini)
 
-if [ "${SCREENSAVER}" = "yes" ]; then
-  if [ "${FBUFDEV}" = "yes" ]; then
-    KILLPID fim
-    [ "${SCREENSAVER_RNDM}" = "yes" ] && RNDMZE="--random" || RNDMZE=""
-    FRAMEBUFFER="${FBDEVICE}" fim --autozoom --quiet --output-device fb ${RNDMZE} --slideshow ${SCREENSAVER_IVAL} /dev/shm/ > /dev/null 2>&1 &
-  else
+if ! [ -e "${TMPDIR}/tmp/screesaver.pid" ]; then
+  if [ "${SCREENSAVER}" = "yes" ]; then
     KILLPID feh
     [ "${SCREENSAVER_RNDM}" = "yes" ] && RNDMZE="--randomize" || RNDMZE=""
-    feh --quiet --fullscreen --auto-zoom ${RNDMZE} --slideshow-delay ${SCREENSAVER_IVAL} /dev/shm/ &
+    feh --quiet --fullscreen --auto-zoom "${RNDMZE}" --slideshow-delay "${SCREENSAVER_IVAL}" "${TMPDIR}"/ & echo $! > "${TMPDIR}/tmp/screesaver.pid"
   fi
 fi